+++ android/src/main/java/com/tailscale/ipn/ui/view/SettingsView.kt
@@ -39,7 +39,6 @@ import androidx.compose.ui.unit.dp
 import com.tailscale.ipn.R
 import com.tailscale.ipn.ui.Links
 import com.tailscale.ipn.ui.model.IpnLocal
-import com.tailscale.ipn.ui.util.Avatar
 import com.tailscale.ipn.ui.util.defaultPaddingModifier
 import com.tailscale.ipn.ui.util.settingsRowModifier
 import com.tailscale.ipn.ui.viewModel.Setting
